Made changes to index.ts, narcissistic numbers logic and test cases as required
