+++ src/karthik/narcissisticNumbers/narcissisticNumbers.ts
@@ -13,11 +13,32 @@ const checkIfNarcissticNumber = (num: number): boolean => {
     sum += parseInt(arr[i]) ** noOfDigits;
   }
 
-  if (num == sum) {
+  if (num === sum) {
     return true;
   } else {
     return false;
   }
 };
 
-export { checkIfNarcissticNumber };
+/**
+ * Function to Find all numbers in range if they are Naricissistic in value
+ * @param {number} from - starting value of the range
+ * @param {number} to - ending value of the range
+ * @return {number[]} - Array of Naricissistic Numbers
+ */
+const getNarcissticNumbersWithinRange = (
+  from: number,
+  to: number
+): Array<number> => {
+  const naricissisticNumbers: Array<number> = [];
+  for (let checkNum = from; checkNum <= to; checkNum++) {
+    const checkIfNaricissistic: boolean = checkIfNarcissticNumber(checkNum);
+    if (checkIfNaricissistic) {
+      naricissisticNumbers.push(checkNum);
+    }
+  }
+
+  return naricissisticNumbers;
+};
+
+export { checkIfNarcissticNumber, getNarcissticNumbersWithinRange };
